+++ src/Client/VeldridRenderer/Rendering/VeldridRenderer.cs
@@ -16,9 +16,12 @@
  */
 
 using System;
+using System.Runtime.InteropServices;
 using Castle.Core.Logging;
+using ImGuiNET;
 using Sovereign.ClientCore.Rendering;
 using Sovereign.ClientCore.Rendering.Configuration;
+using Sovereign.ClientCore.Rendering.Gui;
 using Sovereign.ClientCore.Rendering.Scenes;
 using Sovereign.VeldridRenderer.Rendering.Resources;
 using Veldrid;
@@ -34,6 +37,8 @@ public class VeldridRenderer : IRenderer
     ///     Veldrid graphics device.
     /// </summary>
     private readonly VeldridDevice device;
+
+    private readonly CommonGuiManager guiManager;
 
     /// <summary>
     ///     Veldrid resource manager.
@@ -56,12 +61,13 @@ public class VeldridRenderer : IRenderer
     private bool isDisposed;
 
     public VeldridRenderer(VeldridDevice device, VeldridResourceManager resourceManager,
-        SceneManager sceneManager, VeldridSceneConsumer sceneConsumer)
+        SceneManager sceneManager, VeldridSceneConsumer sceneConsumer, CommonGuiManager guiManager)
     {
         this.device = device;
         this.resourceManager = resourceManager;
         this.sceneManager = sceneManager;
         this.sceneConsumer = sceneConsumer;
+        this.guiManager = guiManager;
     }
 
     public ILogger Logger { private get; set; } = NullLogger.Instance;
@@ -117,6 +123,9 @@ public class VeldridRenderer : IRenderer
             sceneConsumer.ConsumeScene(scene);
             scene.EndScene();
 
+            // GUI rendering from Dear ImGui.
+            RenderGui(commandList);
+
             /* Render and present the next frame. */
             commandList.End();
             device.Device.SubmitCommands(commandList);
@@ -127,5 +136,61 @@ public class VeldridRenderer : IRenderer
         {
             Logger.Error("Error during rendering.", e);
         }
+    }
+
+    /// <summary>
+    ///     Renders the GUI layer.
+    /// </summary>
+    /// <param name="commandList">Active command list.</param>
+    private unsafe void RenderGui(CommandList commandList)
+    {
+        // First stage, convert pending ImGui commands to drawing-level data
+        var drawData = guiManager.Render();
+
+        // Second stage, set up GUI rendering resources.
+        // Check validity.
+        var vertexBuf = resourceManager.GuiVertexBuffer;
+        var indexBuf = resourceManager.GuiIndexBuffer;
+        if (vertexBuf == null || indexBuf == null)
+            throw new InvalidOperationException("Tried to render GUI without buffers.");
+        if (drawData.CmdListsCount == 0) return;
+        if (drawData.TotalVtxCount > vertexBuf.Length)
+            throw new InvalidOperationException("Vertex buffer too small for GUI render.");
+        if (drawData.TotalIdxCount > indexBuf.Length)
+            throw new InvalidOperationException("Index buffer too small for GUI render.");
+
+        // Populate buffers.
+        var vertexOffset = 0;
+        var indexOffset = 0;
+        var vertexSize = Marshal.SizeOf(typeof(ImDrawVert));
+        var indexSize = Marshal.SizeOf<ushort>();
+        for (var i = 0; i < drawData.CmdListsCount; ++i)
+        {
+            var curList = drawData.CmdLists[i];
+
+            // Grab buffers from ImGui.
+            var vertexIn = curList.VtxBuffer.Data.ToPointer();
+            var vertexOut = (vertexBuf.BufferPtr + vertexOffset * vertexSize).ToPointer();
+            var vertexBytes = curList.VtxBuffer.Size * vertexSize;
+            var indexIn = curList.IdxBuffer.Data.ToPointer();
+            var indexOut = (indexBuf.BufferPtr + indexOffset * indexSize).ToPointer();
+            var indexBytes = curList.IdxBuffer.Size * indexSize;
+
+            // Copy into the device buffers. This is a redundant copy but we can optimize it away later
+            // if it's a significant performance impact.
+            Buffer.MemoryCopy(vertexIn, vertexOut,
+                vertexSize * (vertexBuf.Length - vertexOffset),
+                vertexBytes);
+            Buffer.MemoryCopy(indexIn, indexOut,
+                indexSize * (indexBuf.Length - indexOffset),
+                indexBytes);
+
+            vertexOffset += curList.VtxBuffer.Size;
+            indexOffset += curList.IdxBuffer.Size;
+        }
+
+        // Synchronize buffers to device.
+        vertexBuf.Update(commandList);
+        indexBuf.Update(commandList);
     }
 }